+++ lib/commons/comparison_symbol_render/symbol_render_cylinder.dart
@@ -1,0 +1,9 @@
+part of 'symbol_render.dart';
+
+class ComparisonSymbolRenderCylinder extends ComparisonSymbolRender {
+  const ComparisonSymbolRenderCylinder();
+  @override
+  common.PointSymbolRenderer getRenderer() {
+    return common.CylinderSymbolRenderer();
+  }
+}