+++ lib/commons/comparison_symbol_render/symbol_render_rectangle_range.dart
@@ -1,0 +1,9 @@
+part of 'symbol_render.dart';
+
+class ComparisonSymbolRenderRectangleRange extends ComparisonSymbolRender {
+  const ComparisonSymbolRenderRectangleRange();
+  @override
+  common.PointSymbolRenderer getRenderer() {
+    return common.RectangleRangeSymbolRenderer();
+  }
+}