+++ lib/commons/config_render/config_render.dart
@@ -1,6 +1,7 @@
 import 'dart:math';
 
 import 'package:community_charts_common/community_charts_common.dart' as common;
+import '../comparison_symbol_render/symbol_render.dart';
 
 import '../decorator/decorator.dart';
 import '../enums.dart';
@@ -12,4 +13,6 @@ part 'config_render_line.dart';
 part 'config_render_pie.dart';
 part 'config_render_point.dart';
 
-abstract class ConfigRender {}
+abstract class ConfigRender {
+  const ConfigRender();
+}