+++ lib/ordinal/combo.dart
@@ -7,6 +7,7 @@ import 'package:flutter/material.dart';
 
 import '../commons/axis/axis.dart';
 import '../commons/config_render/config_render.dart';
+import '../commons/constants.dart';
 import '../commons/data_model/data_model.dart';
 import '../commons/enums.dart';
 import '../commons/layout_margin.dart';
@@ -44,6 +45,9 @@ class DChartComboO extends StatelessWidget {
 
   /// customize measure axis
   final MeasureAxis? measureAxis;
+
+  /// customize secondary measure axis
+  final MeasureAxis? secondaryMeasureAxis;
 
   /// `areaColor` returns the area color for a given data value.\
   /// If not provided, then group color will be used 10% opacity by default.\
@@ -100,6 +104,7 @@ class DChartComboO extends StatelessWidget {
     this.animationDuration = const Duration(milliseconds: 300),
     this.domainAxis,
     this.measureAxis,
+    this.secondaryMeasureAxis,
     this.areaColor,
     this.fillPattern,
     this.fillColor,
@@ -121,12 +126,16 @@ class DChartComboO extends StatelessWidget {
         OrdinalGroup group = groupList[indexGroup];
         Color groupColor = group.color ??
             Colors.primaries[Random().nextInt(Colors.primaries.length)];
-        return charts.Series<OrdinalData, String>(
+        final chartSeries = charts.Series<OrdinalData, String>(
           id: group.id,
           data: group.data,
           seriesCategory: group.seriesCategory,
           domainFn: (datum, index) => datum.domain,
+          domainLowerBoundFn: (datum, index) => datum.domainLowerBound,
+          domainUpperBoundFn: (datum, index) => datum.domainUpperBound,
           measureFn: (datum, index) => datum.measure,
+          measureLowerBoundFn: (datum, index) => datum.measureLowerBound,
+          measureUpperBoundFn: (datum, index) => datum.measureUpperBound,
           colorFn: (datum, index) => MethodCommon.chartColor(groupColor),
           areaColorFn: areaColor == null
               ? null
@@ -158,6 +167,13 @@ class DChartComboO extends StatelessWidget {
               ? null
               : (datum, index) => barLabelValue!(group, datum, index),
         )..setAttribute(charts.rendererIdKey, group.chartType.name);
+
+        if (group.useSecondaryMeasureAxis) {
+          return chartSeries
+            ..setAttribute(
+                common.measureAxisIdKey, Constants.secondaryMeasureAxisId);
+        }
+        return chartSeries;
       }),
       animate: animate,
       animationDuration: animationDuration,