+++ lib/time/scatter.dart
@@ -7,6 +7,7 @@ import 'package:flutter/material.dart';
 
 import '../commons/axis/axis.dart';
 import '../commons/config_render/config_render.dart';
+import '../commons/constants.dart';
 import '../commons/data_model/data_model.dart';
 import '../commons/enums.dart';
 import '../commons/layout_margin.dart';
@@ -34,6 +35,9 @@ class DChartScatterT extends StatelessWidget {
 
   /// customize measure axis
   final MeasureAxis? measureAxis;
+
+  /// customize secondary measure axis
+  final MeasureAxis? secondaryMeasureAxis;
 
   /// `areaColor` returns the area color for a given data value.\
   /// If not provided, then group color will be used 10% opacity by default.\
@@ -89,6 +93,7 @@ class DChartScatterT extends StatelessWidget {
     this.animationDuration = const Duration(milliseconds: 300),
     this.domainAxis,
     this.measureAxis,
+    this.secondaryMeasureAxis,
     this.areaColor,
     this.fillPattern,
     this.fillColor,
@@ -109,12 +114,16 @@ class DChartScatterT extends StatelessWidget {
         TimeGroup group = groupList[indexGroup];
         Color groupColor = group.color ??
             Colors.primaries[Random().nextInt(Colors.primaries.length)];
-        return charts.Series<TimeData, DateTime>(
+        final chartSeries = charts.Series<TimeData, DateTime>(
           id: group.id,
           data: group.data,
           seriesCategory: group.seriesCategory,
           domainFn: (datum, index) => datum.domain,
+          domainLowerBoundFn: (datum, index) => datum.domainLowerBound,
+          domainUpperBoundFn: (datum, index) => datum.domainUpperBound,
           measureFn: (datum, index) => datum.measure,
+          measureLowerBoundFn: (datum, index) => datum.measureLowerBound,
+          measureUpperBoundFn: (datum, index) => datum.measureUpperBound,
           colorFn: (datum, index) => MethodCommon.chartColor(groupColor),
           areaColorFn: areaColor == null
               ? null
@@ -146,6 +155,12 @@ class DChartScatterT extends StatelessWidget {
               ? null
               : (datum, index) => barLabelValue!(group, datum, index),
         );
+        if (group.useSecondaryMeasureAxis) {
+          return chartSeries
+            ..setAttribute(
+                common.measureAxisIdKey, Constants.secondaryMeasureAxisId);
+        }
+        return chartSeries;
       }),
       flipVerticalAxis: flipVertical,
       animate: animate,
@@ -223,6 +238,49 @@ class DChartScatterT extends StatelessWidget {
                 measureAxis?.tickLabelFormatter,
               ),
               tickProviderSpec: measureAxis?.numericTickProvider?.getRender(),
+            ),
+      secondaryMeasureAxis: secondaryMeasureAxis == null
+          ? null
+          : common.NumericAxisSpec(
+              viewport: secondaryMeasureAxis?.numericViewport?.getRender(),
+              renderSpec: secondaryMeasureAxis?.noRenderSpec ?? false
+                  ? common.NoneRenderSpec(
+                      axisLineStyle:
+                          secondaryMeasureAxis?.lineStyle.getRender(),
+                    )
+                  : secondaryMeasureAxis!.useGridLine
+                      ? common.GridlineRendererSpec(
+                          axisLineStyle:
+                              secondaryMeasureAxis?.lineStyle.getRender(),
+                          lineStyle:
+                              secondaryMeasureAxis?.gridLineStyle.getRender(),
+                          labelStyle:
+                              secondaryMeasureAxis?.labelStyle.getRender(),
+                          labelOffsetFromAxisPx:
+                              secondaryMeasureAxis?.gapAxisToLabel,
+                          labelAnchor: MethodCommon.tickLabelAnchor(
+                              secondaryMeasureAxis?.labelAnchor),
+                          tickLengthPx: secondaryMeasureAxis?.tickLength,
+                        )
+                      : common.SmallTickRendererSpec(
+                          axisLineStyle:
+                              secondaryMeasureAxis?.lineStyle.getRender(),
+                          lineStyle:
+                              secondaryMeasureAxis?.tickLineStyle.getRender(),
+                          labelStyle:
+                              secondaryMeasureAxis?.labelStyle.getRender(),
+                          labelOffsetFromAxisPx:
+                              secondaryMeasureAxis?.gapAxisToLabel,
+                          labelAnchor: MethodCommon.tickLabelAnchor(
+                              secondaryMeasureAxis?.labelAnchor),
+                          tickLengthPx: secondaryMeasureAxis?.tickLength,
+                        ),
+              showAxisLine: secondaryMeasureAxis?.showLine,
+              tickFormatterSpec: common.BasicNumericTickFormatterSpec(
+                secondaryMeasureAxis?.tickLabelFormatter,
+              ),
+              tickProviderSpec:
+                  secondaryMeasureAxis?.numericTickProvider?.getRender(),
             ),
       layoutConfig: layoutMargin?.getRender() ?? LayoutMargin.defaultRender,
       behaviors: [